+++ src/webview/image-manager/components/display-type/index.tsx
@@ -1,10 +1,10 @@
 import { useUpdateEffect } from 'ahooks'
+import { useControlledState } from 'ahooks-x'
 import { Badge, Checkbox, theme } from 'antd'
 import { produce } from 'immer'
 import { difference, uniq } from 'lodash-es'
 import { memo, useMemo } from 'react'
 import { RxViewNone } from 'react-icons/rx'
-import { useControlledState } from 'x-ahooks'
 import GlobalContext, { type RestrictImageFilterType } from '../../contexts/global-context'
 
 export type DisplayTypeFilter = RestrictImageFilterType<{