+++ src/webview/ui-framework/src/components/custom-config-provider/components/primary-color-picker/index.tsx
@@ -1,11 +1,11 @@
 import { TinyColor } from '@ctrl/tinycolor'
 import { useMemoizedFn } from 'ahooks'
+import { useControlledState } from 'ahooks-x'
 import { Button, ColorPicker, type ColorPickerProps } from 'antd'
 import { uniq } from 'lodash-es'
 import { memo, type ReactNode, useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 import { MdOutlineColorLens } from 'react-icons/md'
-import { useControlledState } from 'x-ahooks'
 import { builtInColors, vscodeColors } from '~/webview/ui-framework/src/utils/theme'
 
 type PrimaryColorPickerProps = {